pyproject.toml
[tool.poetry]
name = "michael-django-project-01"
version = "0.1.0"
description = ""
authors = ["Michael Jakobsen"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
dj-rest-auth = "^2.2.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.942"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
